Mobile: Convert the notebook long-press menu to a drawer menu
